Move ModelDownloader to caffe2/python/models
